Separate animations js
