+++ src/css/animations.scss
@@ -56,3 +56,5 @@
     }
   }
 }
+
+/* What we can do for you */
