+++ src/index.js
@@ -1,4 +1,4 @@
 import "./css/index.scss";
 import './scripts/cookieconsent.min';
 import './scripts/script.js';
-import './scripts/factory';
+import './scripts/animations/factory';
