+++ src/scripts/animations/factory.js
